+++ src/project6/QueryNode.java
@@ -21,7 +21,7 @@ public class QueryNode extends absBayesNode implements IBayesNode {
 			count++;
 		}//end for loop
 		//look through CPT to find a true condition in cpt
-		for(int val : cpt){
+		for(double val : cpt){
 			
 		}
 		//value = thisValue;
